src/test/javascript/spec/app/entities/ecom-store-color-theme/ecom-store-color-theme.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { EcomStoreColorThemeService } from 'app/entities/ecom-store-color-theme/ecom-store-color-theme.service';
import { IEcomStoreColorTheme, EcomStoreColorTheme } from 'app/shared/model/ecom-store-color-theme.model';

describe('Service Tests', () => {
  describe('EcomStoreColorTheme Service', () => {
    let injector: TestBed;
    let service: EcomStoreColorThemeService;
    let httpMock: HttpTestingController;
    let elemDefault: IEcomStoreColorTheme;
    let expectedResult: IEcomStoreColorTheme | IEcomStoreColorTheme[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(EcomStoreColorThemeService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new EcomStoreColorTheme(0, 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a EcomStoreColorTheme', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new EcomStoreColorTheme()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a EcomStoreColorTheme', () => {
        const returnedFromService = Object.assign(
          {
            primary: 'BBBBBB',
            secondary: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of EcomStoreColorTheme', () => {
        const returnedFromService = Object.assign(
          {
            primary: 'BBBBBB',
            secondary: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a EcomStoreColorTheme', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
